shuffle around of types
