+++ plutus-chain-index/src/Plutus/ChainIndex/Types.hs
@@ -13,17 +13,23 @@ module Plutus.ChainIndex.Types(
     , Point(..)
     , pointsToTip
     , tipAsPoint
+    , TxValidity(..)
+    , TxStatus(..)
+    , BlockNumber(..)
+    , Depth(..)
     ) where
 
-import           Data.Aeson        (FromJSON, ToJSON)
-import           Data.Default      (Default (..))
-import           Data.Set          (Set)
-import qualified Data.Set          as Set
-import           GHC.Generics      (Generic)
-import           Ledger.Blockchain (BlockId (..))
-import           Ledger.Slot       (Slot)
-import           Numeric.Natural   (Natural)
-import           Prettyprinter     (Pretty (..), (<+>))
+import           Data.Aeson                       (FromJSON, ToJSON)
+import           Data.Default                     (Default (..))
+import           Data.Set                         (Set)
+import qualified Data.Set                         as Set
+import           Data.Text.Prettyprint.Doc.Extras (PrettyShow (..))
+import           GHC.Generics                     (Generic)
+import           Ledger.Blockchain                (BlockId (..))
+import           Ledger.Slot                      (Slot)
+import           Numeric.Natural                  (Natural)
+import           PlutusTx.Lattice                 (MeetSemiLattice (..))
+import           Prettyprinter                    (Pretty (..), (<+>))
 
 newtype PageSize = PageSize { getPageSize :: Natural }
     deriving stock (Eq, Ord, Show, Generic)
@@ -124,3 +130,57 @@ instance Pretty Tip where
         <> ", blockNo="
         <+> pretty tipBlockNo
         <>  ")"
+
+-- | Validity of a transaction that has been added to the ledger
+data TxValidity = TxValid | TxInvalid | UnknownValidity
+  deriving stock (Eq, Ord, Show, Generic)
+  deriving anyclass (ToJSON, FromJSON)
+  deriving Pretty via (PrettyShow TxValidity)
+
+instance MeetSemiLattice TxValidity where
+  TxValid /\ TxValid     = TxValid
+  TxInvalid /\ TxInvalid = TxInvalid
+  _ /\ _                 = UnknownValidity
+
+
+-- | How many blocks deep the tx is on the chain
+newtype Depth = Depth Int
+    deriving stock (Eq, Ord, Show, Generic)
+    deriving newtype (Num, Real, Enum, Integral, Pretty, ToJSON, FromJSON)
+
+instance MeetSemiLattice Depth where
+  Depth a /\ Depth b = Depth (max a b)
+
+{- Note [TxStatus state machine]
+
+The status of a transaction is described by the following state machine.
+
+Current state | Next state(s)
+-----------------------------------------------------
+Unknown       | OnChain
+OnChain       | OnChain, Unknown, Committed
+Committed     | -
+
+The initial state after submitting the transaction is Unknown.
+-}
+
+-- | The status of a Cardano transaction
+data TxStatus =
+  Unknown -- ^ The transaction is not on the chain. That's all we can say.
+  | TentativelyConfirmed Depth TxValidity -- ^ The transaction is on the chain, n blocks deep. It can still be rolled back.
+  | Committed TxValidity -- ^ The transaction is on the chain. It cannot be rolled back anymore.
+  deriving stock (Eq, Ord, Show, Generic)
+  deriving anyclass (ToJSON, FromJSON)
+  deriving Pretty via (PrettyShow TxStatus)
+
+instance MeetSemiLattice TxStatus where
+  Unknown /\ a                                             = a
+  a /\ Unknown                                             = a
+  TentativelyConfirmed d1 v1 /\ TentativelyConfirmed d2 v2 = TentativelyConfirmed (d1 /\ d2) (v1 /\ v2)
+  TentativelyConfirmed _ v1 /\ Committed v2                = Committed (v1 /\ v2)
+  Committed v1 /\ TentativelyConfirmed _ v2                = Committed (v1 /\ v2)
+  Committed v1 /\ Committed v2                             = Committed (v1 /\ v2)
+
+newtype BlockNumber = BlockNumber Int
+    deriving stock (Eq, Ord, Show, Generic)
+    deriving newtype (Num, Real, Enum, Integral, Pretty, ToJSON, FromJSON)
